+++ src/utils/tracker.js
@@ -1,5 +1,5 @@
 import Azk from 'azk';
-import { Q, _, config, log } from 'azk';
+import { Q, _, config, log, t } from 'azk';
 import { meta as azkMeta } from 'azk';
 import { calculateHash } from 'azk/utils';
 
@@ -75,8 +75,8 @@ export class TrackerEvent {
           });
         }
         return tracking_result;
-      }, (error) => {
-        log.error(error.stack);
+      }, () => {
+        log.info(t("tracking.timeout"));
         return false;
       });
   }